Refactored vaults utils
