+++ src/pages/vaults/[vaultid]/_components/VaultIdDetails.tsx
@@ -5,7 +5,7 @@ import { OverflowTable } from '@components/commons/OverflowTable'
 import classNames from 'classnames'
 import BigNumber from 'bignumber.js'
 import React from 'react'
-import { LiquidatedVaultDerivedValues } from '../../../../utils/vaults/LiquidatedVaultDerivedValues'
+import { LiquidatedVaultDerivedValues } from '../../utils/LiquidatedVaultDerivedValues'
 import { TextTruncate } from '@components/commons/text/TextTruncate'
 import { VaultHealthBar } from '../../_components/commons/VaultHealthBar'
 import { VaultNumberValues } from '../../_components/commons/VaultNumberValues'
