+++ src/pages/vaults/_components/VaultMobileCard.tsx
@@ -4,7 +4,7 @@ import React, { useState } from 'react'
 import { Transition } from '@headlessui/react'
 import BigNumber from 'bignumber.js'
 import ReactNumberFormat from 'react-number-format'
-import { LiquidatedVaultDerivedValues } from '../../../utils/vaults/LiquidatedVaultDerivedValues'
+import { LiquidatedVaultDerivedValues } from '../utils/LiquidatedVaultDerivedValues'
 import { TextTruncate } from '@components/commons/text/TextTruncate'
 import { VaultLink } from '@components/commons/link/VaultLink'
 import { VaultStatus } from './commons/VaultStatus'
