+++ src/pages/vaults/_components/commons/VaultAuctionDetails.tsx
@@ -3,7 +3,7 @@ import { Link } from '@components/commons/link/Link'
 import ReactNumberFormat from 'react-number-format'
 import BigNumber from 'bignumber.js'
 import { VaultTokenSymbols } from './VaultTokenSymbols'
-import { calculateCollateralsValue } from '../../../../utils/vaults/CalculateCollateralsValue'
+import { calculateCollateralsValue } from '../../utils/CalculateCollateralsValue'
 import { LoanVaultLiquidated, LoanVaultLiquidationBatch } from '@defichain/whale-api-client/dist/api/loan'
 import { getAssetIcon } from '@components/icons/assets/tokens'
 import { VaultDetailsListItem } from './VaultDetailsListItem'
