+++ src/pages/vaults/utils/CalculateCollateralsValue.tsx
