+++ src/pages/vaults/utils/NextCollateralizationRatio.tsx
